+++ src/components/storage/index.ts
@@ -1,0 +1,2 @@
+export { default as StorageStatus } from './StorageStatus';
+export { StorageProvider, useStorageContext } from './StorageProvider';